@@ -1,12 +1,13 @@
 module github.com/goal-web/supports
 
-go 1.17
+go 1.18
 
 require (
 	github.com/apex/log v1.9.0
 	github.com/goal-web/contracts v0.1.46
 	github.com/modood/table v0.0.0-20200225102042-88de94bb9876
 	github.com/pkg/errors v0.8.1
+	github.com/spf13/cast v1.4.1
 	github.com/stretchr/testify v1.7.0
 )
 
@@ -16,3 +17,5 @@ require (
 	github.com/smartystreets/goconvey v1.7.2 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200605160147-a5ece683394c // indirect
 )
+
+replace github.com/goal-web/contracts => ../contracts
