@@ -3,10 +3,12 @@ package supports
 import (
 	"github.com/goal-web/contracts"
 	"github.com/goal-web/supports/utils"
+	"github.com/spf13/cast"
 	"strings"
+	"time"
 )
 
-type InstanceGetter func(key string) interface{}
+type InstanceGetter func(key string) any
 
 type BaseFields struct { // 具体方法
 	contracts.FieldsProvider // 抽象方法，继承 interface
@@ -14,7 +16,141 @@ type BaseFields struct { // 具体方法
 	Getter InstanceGetter // 如果有设置 getter ，优先使用 getter
 }
 
-func (this *BaseFields) get(key string) interface{} {
+func (this *BaseFields) Optional(key string, defaultValue any) any {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalTime(key string, defaultValue time.Time) time.Time {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalDuration(key string, defaultValue time.Duration) time.Duration {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalInt8(key string, defaultValue int8) int8 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalInt16(key string, defaultValue int16) int16 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalInt32(key string, defaultValue int32) int32 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalInt64(key string, defaultValue int64) int64 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalUint(key string, defaultValue uint) uint {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalUint8(key string, defaultValue uint8) uint8 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalUint16(key string, defaultValue uint16) uint16 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalUint32(key string, defaultValue uint32) uint32 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalUint64(key string, defaultValue uint64) uint64 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalFloat64(key string, defaultValue float64) float64 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalFloat32(key string, defaultValue float32) float32 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) OptionalBool(key string, defaultValue bool) bool {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) GetTime(key string) time.Time {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) GetDuration(key string) time.Duration {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) GetInt8(key string) int8 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) GetInt16(key string) int16 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) GetInt32(key string) int32 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) GetUint(key string) uint {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) GetUint8(key string) uint8 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) GetUint16(key string) uint16 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) GetUint32(key string) uint32 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) GetUint64(key string) uint64 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) GetFloat32(key string) float32 {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (this *BaseFields) Get(key string) any {
+
+}
+
+func (this *BaseFields) get(key string) any {
 	if this.Getter != nil {
 		if value := this.Getter(key); value != nil && value != "" {
 			return value
@@ -62,49 +198,24 @@ func (this *BaseFields) OnlyExists(keys ...string) contracts.Fields {
 	return fields
 }
 
-func (this *BaseFields) StringOption(key string, defaultValue string) string {
+func (this *BaseFields) OptionalString(key string, defaultValue string) string {
 	if value := this.get(key); value != nil && value != "" {
 		return utils.ConvertToString(value, defaultValue)
 	}
 	return defaultValue
 }
 
-func (this *BaseFields) Int64Option(key string, defaultValue int64) int64 {
+func (this *BaseFields) OptionalInt(key string, defaultValue int) int {
 	if value := this.get(key); value != nil && value != "" {
-		return utils.ConvertToInt64(value, defaultValue)
+		if target, err := cast.ToIntE(value); err == nil {
+			return target
+		}
+		return defaultValue
 	}
 	return defaultValue
 }
 
-func (this *BaseFields) IntOption(key string, defaultValue int) int {
-	if value := this.get(key); value != nil && value != "" {
-		return utils.ConvertToInt(value, defaultValue)
-	}
-	return defaultValue
-}
-
-func (this *BaseFields) Float64Option(key string, defaultValue float64) float64 {
-	if value := this.get(key); value != nil && value != "" {
-		return utils.ConvertToFloat64(value, defaultValue)
-	}
-	return defaultValue
-}
-
-func (this *BaseFields) FloatOption(key string, defaultValue float32) float32 {
-	if value := this.get(key); value != nil && value != "" {
-		return utils.ConvertToFloat(value, defaultValue)
-	}
-	return defaultValue
-}
-
-func (this *BaseFields) BoolOption(key string, defaultValue bool) bool {
-	if value := this.get(key); value != nil && value != "" {
-		return utils.ConvertToBool(value, defaultValue)
-	}
-	return defaultValue
-}
-
-func (this *BaseFields) FieldsOption(key string, defaultValue contracts.Fields) contracts.Fields {
+func (this *BaseFields) OptionalFields(key string, defaultValue contracts.Fields) contracts.Fields {
 	if value := this.get(key); value != nil && value != "" {
 		if fields, err := utils.ConvertToFields(value); err == nil {
 			return fields
@@ -123,29 +234,29 @@ func (this *BaseFields) FieldsOption(key string, defaultValue contracts.Fields) 
 }
 
 func (this *BaseFields) GetString(key string) string {
-	return this.StringOption(key, "")
+	return this.OptionalString(key, "")
 }
 
 func (this *BaseFields) GetInt64(key string) int64 {
-	return this.Int64Option(key, 0)
+	return this.OptionalInt64(key, 0)
 }
 
 func (this *BaseFields) GetInt(key string) int {
-	return this.IntOption(key, 0)
+	return this.OptionalInt(key, 0)
 }
 
 func (this *BaseFields) GetFloat64(key string) float64 {
-	return this.Float64Option(key, 0)
+	return this.OptionalFloat64(key, 0)
 }
 
 func (this *BaseFields) GetFloat(key string) float32 {
-	return this.FloatOption(key, 0)
+	return this.OptionalFloat32(key, 0)
 }
 
 func (this *BaseFields) GetBool(key string) bool {
-	return this.BoolOption(key, false)
+	return this.OptionalBool(key, false)
 }
 
 func (this *BaseFields) GetFields(key string) contracts.Fields {
-	return this.FieldsOption(key, contracts.Fields{})
+	return this.OptionalFields(key, contracts.Fields{})
 }
